remove code duplicates
