+++ src/monitoring/io.py
@@ -4,8 +4,53 @@ Provides a robust JSONL reader that gracefully handles partial or
 malformed lines, making it safe to call against an in-progress log.
 """
 
+import html as html_module
 import json
 import logging
+import os
+
+
+def _escape(text) -> str:
+    """HTML-escape a value, returning <em>N/A</em> for None."""
+    if text is None:
+        return "<em>N/A</em>"
+    return html_module.escape(str(text))
+
+
+def _resolve_image_path(raw_path: str | None, image_base_dir: str | None) -> str | None:
+    """Resolve an image path to an absolute existing path, or None if missing."""
+    if not raw_path:
+        return None
+    if not os.path.isabs(raw_path) and image_base_dir:
+        raw_path = os.path.join(image_base_dir, raw_path)
+    if os.path.isfile(raw_path):
+        return raw_path
+    return None
+
+
+def _build_record_entry(rec: dict) -> dict:
+    """Extract the common sidebar fields shared by all monitoring views."""
+    return {
+        "iteration": rec.get("iteration_number"),
+        "island": rec.get("birth_island"),
+        "batch": rec.get("batch_index"),
+        "train_loss": rec.get("train_loss"),
+        "initial_loss": rec.get("initial_loss"),
+        "n_params": rec.get("n_params"),
+        "complexity_penalty": rec.get("complexity_penalty"),
+        "mode": rec.get("mode"),
+        "llm_name": rec.get("llm_name"),
+        "temperature": rec.get("temperature"),
+        "model_code": rec.get("model_code_numpy"),
+        "param_est_code": rec.get("param_est_code"),
+        "model_prompt": rec.get("model_prompt"),
+        "image_prompt_path": rec.get("image_prompt_path"),
+        "model_llm_response": rec.get("model_llm_response"),
+        "param_est_prompt": rec.get("param_est_prompt"),
+        "param_est_llm_response": rec.get("param_est_llm_response"),
+        "train_fit_image_path": rec.get("train_fit_image_path"),
+        "test_fit_image_path": rec.get("test_fit_image_path"),
+    }
 
 
 def load_generation_log(log_path: str) -> list[dict]:
